Add errorMessage property, Add check and ask for confirmation for delete method

+++ HogwartsPotionsFrontend/src/app/all-rooms/all-rooms.component.ts
@@ -15,6 +15,7 @@ export class AllRoomsComponent implements OnInit {
   url = "/api/room";
   rooms: Room[] = [];
   houseTypes: String[] = ["Gryffindor", "Hufflepuff", "Ravenclaw", "Slytherin"];
+  errorMessage: string | undefined;
 
   ngOnInit(): void {
     this.getRooms();
@@ -28,7 +29,14 @@ export class AllRoomsComponent implements OnInit {
   }
 
   delete(room: Room): void {
-    this.rooms = this.rooms.filter(r => r !== room);
-    this.roomService.deleteRoom(room.id).subscribe();
+    if (room.residents.length > 0) {
+      alert("You can delete only empty rooms!");
+      return;
+    }
+    
+    if (confirm("Are you sure you want to delete this room? \n This will permanetle remove this room from Hogwarts!")){
+      this.rooms = this.rooms.filter(r => r !== room);
+      this.roomService.deleteRoom(room.id).subscribe();
+    }
   }
 }
